+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Mapbox welcomes participation and contributions from everyone.
 
 ## 11.27.0 - 24 July, 2026
 
+* Fixed a crash caused by a data race when subscribing to or cancelling a `Signal` (such as `onCameraChanged`) from a background thread while it emits on another thread. Protected other `Signal`s from the same scenario.
+
 ## 11.27.0-rc.1 - 13 July, 2026
 
 ## 11.26.0 - 09 July, 2026